Move test_align_shape_to_tile to test_tensor_utils.py
align_shape_to_tile lives in models/common/tensor_utils.py, so its test
belongs with the rest of that module's tests in
models/common/tests/modules/test_tensor_utils.py rather than in
models/common/tests/test_utils.py, which covers models/common/tests/utils.py.

Also converts two pre-existing pytest.raises usages in that file to the
expect_error fixture to satisfy the prefer-expect-error pre-commit hook,
which scans the whole file whenever it's touched.
